refactor(docs): #704 PR-B — extract _shared + _status_tracker submodules
First extraction in the routes/__init__.py shrink chain (#704). Moves
the cross-cutting helpers (status badge, timestamp formatter, stale
checks, polling-window helpers, elapsed-time formatters, pipeline
constants) to _shared.py and the big 6-stage tracker renderer to
_status_tracker.py. __init__.py drops ~400 lines (3015 → 2613) and
re-exports every moved name via __all__ so existing test patches and
imports keep working without any patch-path swap. PR-C/D/E will move
the remaining handler bodies (_list, _upload, _detail) and bring the
package under the 400-line cap.
